# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ForgeMind AI - Industrial Knowledge Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signUp
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignUpInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenInput"
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Reset email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List all documents
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Document list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentListResponse"
...
  /chat/sessions:
    get:
      operationId: listChatSessions
      tags: [chat]
      summary: List all chat sessions
      responses:
        "200":
          description: Chat sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatSession"
    post:
      operationId: createChatSession
      tags: [chat]
      summary: Create a new chat session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatSessionInput"
      responses:
        "201":
          description: Chat session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSession"
  /chat/sessions/{id}:
    get:
      operationId: getChatSession
      tags: [chat]
      summary: Get a chat session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Chat session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSession"
    delete:
      operationId: deleteChatSession
      tags: [chat]
      summary: Delete a chat session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /chat/sessions/{id}/messages:
    get:
      operationId: getChatMessages
      tags: [chat]
      summary: Get messages in a chat session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message to the AI
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
  /chat/sessions/{id}/export:
    post:
      operationId: exportChatSession
      tags: [chat]
      summary: Export chat session to text/PDF
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Export URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportResponse"
...
  /admin/users/{id}:
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      summary: Admin - delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── AUTH SCHEMAS ─────────────────────────────────────────────────────────
    SignUpInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string

        password:
          type: string
          minLength: 6
        company:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string

        password:
          type: string
    RefreshTokenInput:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    ForgotPasswordInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
    AuthResponse:
      type: object
      required: [access_token, user]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, email, name, role, created_at]
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        company:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, user]
        avatar_url:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time

    # ── DASHBOARD SCHEMAS ─────────────────────────────────────────────────────
...
    Document:
      type: object
      required: [id, filename, file_type, size, status, created_at, category]
      properties:
        id:
          type: string
        filename:
          type: string
        original_filename:
          type: string
        file_type:
          type: string
          enum: [pdf, docx, txt, xlsx, csv]
        size:
          type: integer
        status:
          type: string
          enum: [processing, ready, error, ocr_needed]
        category:
          type: ["string", "null"]
        extracted_text_preview:
          type: ["string", "null"]
        page_count:
          type: ["integer", "null"]
        machine_ids:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        processed_at:
          type: ["string", "null"]
          format: date-time
...
    DocumentListResponse:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Document"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    ChatSession:
      type: object
      required: [id, title, created_at, message_count]
      properties:
        id:
          type: string
        title:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        message_count:
          type: integer
        last_message:
          type: ["string", "null"]
    ChatSessionInput:
      type: object
      properties:
        title:
          type: string
    ChatMessage:
      type: object
      required: [id, session_id, role, content, created_at]
      properties:
        id:
          type: string
        session_id:
          type: string
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        sources:
          type: array
          items:
            $ref: "#/components/schemas/MessageSource"
        confidence_score:
          type: ["number", "null"]
        created_at:
          type: string
          format: date-time
    MessageSource:
      type: object
      required: [document_id, filename, excerpt]
      properties:
        document_id:
          type: string
        filename:
          type: string
        excerpt:
          type: string
        page_number:
          type: ["integer", "null"]
    ChatMessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        language:
          type: string
          enum: [en, hi]
          default: en
    ExportResponse:
      type: object
      required: [export_text]
      properties:
        export_text:
          type: string
        filename:
          type: string

    # ── KNOWLEDGE GRAPH SCHEMAS ──────────────────────────────────────────────
...