# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * ForgeMind AI - Industrial Knowledge Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: string;
  email: string;
  name: string;
  /** @nullable */
  company?: string | null;
  role: UserRole;
  /** @nullable */
  avatar_url?: string | null;
  created_at: Date;
}